Update monthly.js
first of all, really thanks to you.
because i want a responsive event calendar which was fullcalendar not support this.

I amend some of the code because i have meet some issues, may anyone have the same problems.
I am using asp.net MVC, the web page to display this event calendar contains other components. so i want to centralize the ajax call in single function.

mode: eventByJson
jsonObj: {}

eventByJson mode specify to create the event calendar by the json object, there are no ajax action.

A month is a base unit to display event and get xml. when i pass the next month json in it, it will display the next month.
not display the month of new Date();

calendarPeriod: dateObj

use to display a specify month